Move test classes into their own files.
