+++ tests/TestCase/Error/ExceptionRendererTest.php
@@ -942,44 +942,3 @@ class ExceptionRendererTest extends TestCase
         $this->assertContains("'seven' => (int) 7", $result);
     }
 }
-
-class TestErrorController extends Controller
-{
-    /**
-     * uses property
-     *
-     * @var array
-     */
-    public $uses = [];
-
-    /**
-     * components property
-     *
-     * @return void
-     */
-    public $components = ['Blueberry'];
-
-    /**
-     * beforeRender method
-     *
-     * @return \Cake\Http\Response|null
-     */
-    public function beforeRender(EventInterface $event): ?Response
-    {
-        echo $this->Blueberry->testName;
-
-        return null;
-    }
-
-    /**
-     * index method
-     *
-     * @return array
-     */
-    public function index()
-    {
-        $this->autoRender = false;
-
-        return 'what up';
-    }
-}
