+++ tests/test_app/TestApp/Controller/ControllerTestAppController.php
@@ -1,0 +1,18 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\Controller;
+
+use Cake\Controller\Controller;
+
+/**
+ * AppController class
+ */
+class ControllerTestAppController extends Controller
+{
+    /**
+     * modelClass property
+     *
+     * @var string
+     */
+    public $modelClass = 'Posts';
+}
